requirements.txt
Django==1.11.4
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-filter==1.0.4
djangorestframework==3.6.3
djangorestframework-filters==0.10.1
pytz==2017.2
